Use xxhash to cache info about files
Every time a new file is encountered we store its hash in the cache
directory. This is faster than using `file` every single time.  We only
computer the hash of the first 100000 bytes of each file, so it is a
fast process even for larger files like videos.

Currently we store info about images, videos and other files that should
not be shown by ucollage. This technique permits `ucollage` to utilize
`file` as little as possible.

Using hashing also permits to maintain the relevant information even if
the file is moved to another location.